requirements.txt
python-chess
flask>=3.0
